Connected the Home view to the new backend.
There is now "stats-api" up to "dashboard-front-end" continuity in the data.

I've broken a fair bit of stuff in the dashboard doing this. Sorry guys.
Altough, there's a couple of thing that might not be possible to repair.

For example the right graphic, I think was presenting individual data points? If so, it can't be like that. There's also a "time picker". Although great, could lead to privacy issue as people "scrub" this slider a day at a time and note modification to means.

There is also a slight issue of consistency as the back-button is used, I didn't have time to fix it.